+++ collector-exoplayer/src/main/java/com/bitmovin/analytics/exoplayer/ExoUtil.kt
@@ -1,5 +1,7 @@
 package com.bitmovin.analytics.exoplayer
 
+import android.os.Handler
+import android.os.Looper
 import com.google.android.exoplayer2.ExoPlayerLibraryInfo
 import com.google.android.exoplayer2.Player
 
@@ -24,4 +26,20 @@ internal object ExoUtil {
             }
             return "unknown"
         }
+
+    // Method that can be used to make sure a certain code block is executed
+    // on same thread as provided looper
+    // Be careful, given code is either executed synchronously when calling
+    // thread is same thread as applicationLooper, or asynchronously if not
+    // This means code calling this cannot rely on order of execution
+    fun executeSyncOrAsyncOnLooperThread(applicationLooper: Looper, function: () -> Unit) {
+        if (Thread.currentThread() != applicationLooper.thread) {
+            val handler = Handler(applicationLooper)
+            handler.post {
+                function.invoke()
+            }
+        } else {
+            function.invoke()
+        }
+    }
 }